requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
twilio==9.3.2
openai==1.47.1
httpx==0.27.2
pytest==8.3.3
